cleaned up classes and removed highlighter class
