--- highlighter.cpp
@@ -1,37 +1,0 @@
-#include "highlighter.h"
-
-Highlighter::Highlighter(QTextDocument *parent)
-    : QSyntaxHighlighter(parent)
-{
-    HighlightingRule rule;
-
-    keywordFormat.setForeground(Qt::green);
-    keywordFormat.setFontWeight(QFont::Bold);
-    QStringList keywordPatterns;
-    keywordPatterns << "\\binput\\b" << "\\blet\\b"
-                    << "\\blet\\b" << "\\bprint\\b" << "\\bgoto\\b"
-                    << "\\bif\\b" << "\\bend\\b";
-
-    foreach (const QString &pattern, keywordPatterns) {
-        rule.pattern = QRegularExpression(pattern);
-        rule.format = keywordFormat;
-        highlightingRules.append(rule);
-    }
-
-    singleLineCommentFormat.setForeground(Qt::red);
-    rule.pattern = QRegularExpression("rem");
-    rule.format = singleLineCommentFormat;
-    highlightingRules.append(rule);
-}
-
-void Highlighter::highlightBlock(const QString &text)
-{
-    foreach (const HighlightingRule &rule, highlightingRules) {
-        QRegularExpressionMatchIterator matchIterator = rule.pattern.globalMatch(text);
-        while (matchIterator.hasNext()) {
-            QRegularExpressionMatch match = matchIterator.next();
-            setFormat(match.capturedStart(), match.capturedLength(), rule.format);
-        }
-    }
-    setCurrentBlockState(0);
-}
